pi-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pyzbar==0.1.9
pywebview==5.4
python-multipart==0.0.20
qrcode[pil]==7.4.2
